backend/requirements.txt
# Dependencies principales - Compatible Python 3.7
Flask==2.2.5
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Pillow==9.5.0
python-dotenv==0.21.1
gunicorn==20.1.0

# Base de données
psycopg2-binary==2.9.6
SQLAlchemy==1.4.47

# Validation et sérialisation
marshmallow==3.19.0
email-validator==1.3.1

# Cache et sessions
redis==4.5.5
Flask-Session==0.4.1

# Sécurité
bcrypt==4.0.1
cryptography==39.0.1

# Paiement
stripe==5.4.0

# Emailing
requests==2.28.2

# Gestion de fichiers
boto3==1.26.137

# Utilitaires
pytz==2023.3
python-slugify==8.0.1

# Tests
pytest==7.4.3
pytest-flask==1.2.0
pytest-cov==4.1.0

# Développement
black==22.3.0
flake8==5.0.4
